requirements.txt
GitPython==2.1.12
PyYAML==5.1.1